refactoring chord quiz to improve readability
